refactor: move deriv-chart code
